# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TabNorder POS System API",
    "description": "Comprehensive Point of Sale (POS) system API for cafes and restaurants with order management, kitchen operations, cashier functions, and business analytics.",
    "version": "1.0.0",
    "contact": {
      "name": "TabNorder API Support",
      "email": "support@tabnorder.com"
    }
  },
  "servers": [
    {
      "url": "https://us-central1-tab-n-order.cloudfunctions.net",
      "description": "Production server"
    },
    {
      "url": "http://localhost:5001/tab-n-order/us-central1",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/attachments": {
      "post": {
        "tags": ["Attachments"],
        "summary": "Upload attachment",
        "description": "Uploads a file attachment to Firebase Storage and returns the attachment URL. Accepts base64 encoded file in JSON body.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadAttachmentRequest"
              },
              "example": {
                "file": "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mNk+M9QDwADhgGAWjR9awAAAABJRU5ErkJggg==",
                "fileName": "example.png",
                "mimeType": "image/png",
                "uploadedBy": "user123",
                "relatedEntityType": "item",
                "relatedEntityId": "item456",
                "folder": "attachments",
                "metadata": {
                  "description": "Item image"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Attachment uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Attachment"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing or invalid file data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error - file type or size not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Attachments"],
        "summary": "Get attachments with filters",
        "description": "Retrieves attachments with optional filtering using query parameters",
        "parameters": [
          {
            "name": "uploaded_by",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by user ID who uploaded the file",
            "example": "user123"
          },
          {
            "name": "related_entity_type",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by related entity type (e.g., 'item', 'menu', 'place')",
            "example": "item"
          },
          {
            "name": "related_entity_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by related entity ID",
            "example": "item456"
          },
          {
            "name": "file_type",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by file extension (e.g., 'png', 'jpg', 'pdf')",
            "example": "png"
          },
          {
            "name": "mime_type",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by MIME type",
            "example": "image/png"
          }
        ],
        "responses": {
          "200": {
            "description": "Attachments retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Attachment"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/attachments/{id}": {
      "get": {
        "tags": ["Attachments"],
        "summary": "Get attachment by ID",
        "description": "Retrieves a specific attachment by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Attachment ID",
            "example": "attachment123"
          }
        ],
        "responses": {
          "200": {
            "description": "Attachment retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Attachment"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing attachment ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Attachment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Attachments"],
        "summary": "Delete attachment",
        "description": "Deletes an attachment from Firebase Storage and removes its metadata",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Attachment ID",
            "example": "attachment123"
          }
        ],
        "responses": {
          "200": {
            "description": "Attachment deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing attachment ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Attachment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "version": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Attachment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique attachment ID",
            "example": "attachment123"
          },
          "fileName": {
            "type": "string",
            "description": "Generated unique file name",
            "example": "example_a1b2c3d4.png"
          },
          "originalFileName": {
            "type": "string",
            "description": "Original file name from upload",
            "example": "example.png"
          },
          "fileType": {
            "type": "string",
            "description": "File extension without dot",
            "example": "png"
          },
          "mimeType": {
            "type": "string",
            "description": "MIME type of the file",
            "example": "image/png"
          },
          "fileSize": {
            "type": "integer",
            "description": "File size in bytes",
            "example": 1024
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Public URL to access the attachment",
            "example": "https://storage.googleapis.com/bucket/attachments/example_a1b2c3d4.png"
          },
          "storagePath": {
            "type": "string",
            "description": "Path in Firebase Storage",
            "example": "attachments/example_a1b2c3d4.png"
          },
          "uploadedBy": {
            "type": "string",
            "description": "User ID who uploaded the file",
            "example": "user123"
          },
          "relatedEntityType": {
            "type": "string",
            "description": "Type of related entity (e.g., 'item', 'menu', 'place')",
            "example": "item"
          },
          "relatedEntityId": {
            "type": "string",
            "description": "ID of the related entity",
            "example": "item456"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata",
            "additionalProperties": true,
            "example": {
              "description": "Item image"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": ["id", "fileName", "originalFileName", "fileType", "mimeType", "fileSize", "url", "storagePath", "createdAt", "updatedAt"]
      },
      "UploadAttachmentRequest": {
        "type": "object",
        "required": ["file", "fileName", "mimeType"],
        "properties": {
          "file": {
            "type": "string",
            "format": "byte",
            "description": "Base64 encoded file content or data URL (e.g., 'data:image/png;base64,...')",
            "example": "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mNk+M9QDwADhgGAWjR9awAAAABJRU5ErkJggg=="
          },
          "fileName": {
            "type": "string",
            "description": "Original file name",
            "example": "example.png"
          },
          "mimeType": {
            "type": "string",
            "description": "MIME type of the file (e.g., 'image/png', 'image/jpeg', 'application/pdf')",
            "example": "image/png",
            "enum": [
              "image/jpeg",
              "image/jpg",
              "image/png",
              "image/gif",
              "image/webp",
              "image/svg+xml",
              "application/pdf",
              "application/msword",
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
              "application/vnd.ms-excel",
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
            ]
          },
          "uploadedBy": {
            "type": "string",
            "description": "User ID who uploaded the file (optional)",
            "example": "user123"
          },
          "relatedEntityType": {
            "type": "string",
            "description": "Type of related entity (optional, e.g., 'item', 'menu', 'place')",
            "example": "item"
          },
          "relatedEntityId": {
            "type": "string",
            "description": "ID of the related entity (optional)",
            "example": "item456"
          },
          "folder": {
            "type": "string",
            "description": "Folder path in storage (optional, default: 'attachments')",
            "example": "attachments"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata (optional)",
            "additionalProperties": true,
            "example": {
              "description": "Item image"
            }
          }
        }
      }
    }
  },
...
}